% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompDA.R
\name{CompDA}
\alias{CompDA}
\title{Compositional Differential Abundance Analysis for Microbiome Data}
\usage{
CompDA(
  x,
  y,
  covariates = NULL,
  epsilon = NULL,
  family_y = "gaussian",
  method_y = "glmnet",
  cv_folds = 5,
  m = 10000,
  verbose = TRUE,
  debug_dir = NULL
)
}
\arguments{
\item{x}{A matrix of microbiome relative abundances. Rows are samples and columns are microbes.}

\item{y}{The outcome of interests. Currently only continuous and binary y are supported.}

\item{covariates}{Matrix of potential confounders.}

\item{epsilon}{Pseudo count to add to relative abundances before transformation.}

\item{family_y}{Specify the glm family for y. Only "binomial" (for binary y) and "gaussian" (for continuous y) are supported.}

\item{method_y}{Regression model for y. Currently options "glmnet", "xgbTree", "rf", and "svm" are supported.}

\item{cv_folds}{Number of CV folds in the modeling for y.}

\item{m}{Number of random CRT samples to generate for computing p-values.}

\item{verbose}{Whether or not to print verbose information.}

\item{debug_dir}{Directory to save intermediate output (for debugging purpose).}
}
\value{
A matrix of two columns. Each row corresponds to one microbe's testing results. First column indicates conditional associations with the outcome. Second column indicates p-values.
}
\description{
Compositional Differential Abundance Analysis for Microbiome Data
}
